[refactor] 패키지 명 수정
