+++ src/test/java/calculator/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package StringAddCalculator;
+package calculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
